backend/pyproject.toml
[project]
name = "minis-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "cryptography>=46.0.4",
    "fastapi[standard]>=0.128.6",
    "httpx>=0.28.1",
    "langfuse>=2.0.0",
    "litellm>=1.81.9",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.46",
    "sse-starlette>=3.2.0",
    "typer>=0.15.0",
    "trafilatura>=1.12.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
